refactor: repackage http response class
