+++ src/main/java/webserver/RequestHandler.java
@@ -1,7 +1,7 @@
 package webserver;
 
 import http.HttpRequest;
-import http.HttpResponse;
+import http.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import webserver.processor.Processors;
