+++ src/main/java/webserver/processor/NotFoundProcessor.java
@@ -1,7 +1,7 @@
 package webserver.processor;
 
 import http.HttpRequest;
-import http.HttpResponse;
+import http.response.HttpResponse;
 import http.StatusCode;
 
 public class NotFoundProcessor implements Processor {
