+++ src/main/java/webserver/processor/ResourceProcessor.java
@@ -2,7 +2,7 @@ package webserver.processor;
 
 import http.ContentType;
 import http.HttpRequest;
-import http.HttpResponse;
+import http.response.HttpResponse;
 import http.StatusCode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
