+++ src/test/java/webserver/processor/ProcessorsTest.java
@@ -1,7 +1,7 @@
 package webserver.processor;
 
 import http.HttpRequest;
-import http.HttpResponse;
+import http.response.HttpResponse;
 import http.StatusCode;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
